refactor: 가견적서 생성 사용자 입력 관련 리팩토링 (Step4 제거)
