+++ apps/web/src/app/(main)/estimate/_components/ProcessCheckbox/ProcessCheckbox.tsx
@@ -5,9 +5,10 @@ interface ProcessCheckboxProps {
   process: ProcessInfo<ProcessId>;
   selected: boolean;
   onToggle: (id: ProcessId) => void;
+  hasDetail?: boolean;
 }
 
-export const ProcessCheckbox = ({ process, selected, onToggle }: ProcessCheckboxProps) => {
+export const ProcessCheckbox = ({ process, selected, onToggle, hasDetail = false }: ProcessCheckboxProps) => {
   return (
     <button
       type="button"
@@ -15,6 +16,9 @@ export const ProcessCheckbox = ({ process, selected, onToggle }: ProcessCheckbox
       onClick={() => onToggle(process.id)}
       aria-pressed={selected}
     >
+      {hasDetail && (
+        <span className={selected ? styles.detailDotActive : styles.detailDot} title="세부 입력 항목이 있어요" />
+      )}
       <span className={selected ? styles.processNameSelected : styles.processNameUnselected}>{process.name}</span>
       <span className={styles.processDesc}>{process.description}</span>
     </button>
